Added discharge_per_arm function to collector well
We may want to discuss whether we should put this functionality simply in the discharge function (which gives the discharge from each layer for the entire collector well) but add a keyword argument to get the per-arm discharge. If we leave it as is, then the next question is whether discharge_per_arm is a good name.